% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSdists.R
\name{DTNearest_aY}
\alias{DTNearest_aY}
\title{Function uses data.table rolling join to identify the nearest features
really quickly.}
\usage{
DTNearest_aY(DT1, DT2)
}
\arguments{
\item{DT1}{A data.table object to be joined to a second data.table object.}

\item{DT2}{A second data.table object to join with DT1.}
}
\value{
A rolling joined data.table object.
}
\description{
Function uses data.table rolling join to identify the nearest features
really quickly.
}
